requirements.txt
streamlit
google-generativeai
python-dotenv
# pyttsx3
# SpeechRecognition